+++ mpire/pool.py
@@ -130,15 +130,13 @@ class WorkerPool:
         """
         self.pool_params.keep_alive = keep_alive
 
-    def _start_workers(self, progress_bar: bool) -> None:
+    def _start_workers(self) -> None:
         """
         Spawns the workers and starts them so they're ready to start reading from the tasks queue.
-
-        :param progress_bar: Whether there's a progress bar
         """
         # Init communication primitives
         logger.debug("Initializing comms")
-        self._worker_comms.init_comms(self.map_params.worker_exit is not None, progress_bar)
+        self._worker_comms.init_comms()
         self._worker_insights.reset_insights(self.pool_params.enable_insights)
         self._exit_results = []
 
@@ -161,7 +159,7 @@ class WorkerPool:
         for worker_id in self._worker_comms.get_worker_restarts():
             # Obtain results from exit results queue (should be done before joining the worker)
             if self.map_params.worker_exit:
-                self._exit_results.append(self._worker_comms.get_exit_results(worker_id))
+                self._get_exit_result_from_worker(worker_id)
 
             # Join worker
             self._worker_comms.reset_worker_restart(worker_id)
@@ -190,10 +188,25 @@ class WorkerPool:
                 worker_died = self._worker_comms.is_worker_alive(worker_id) and not self._workers[worker_id].is_alive()
             if worker_died:
                 # We need to add an exception if we're using the progress bar handler
-                if self._worker_comms.has_progress_bar():
-                    self._worker_comms.add_exception(RuntimeError, f"Worker-{worker_id} died unexpectedly")
+                err_msg = f"Worker-{worker_id} died unexpectedly"
+                if self.map_params.progress_bar:
+                    self._worker_comms.add_exception(RuntimeError, err_msg)
                 self.terminate()
-                raise RuntimeError(f"Worker-{worker_id} died unexpectedly")
+                raise RuntimeError(err_msg)
+
+        # Check for worker_init/task/worker_exit timeouts
+        for timeout_var, has_timed_out_func, timeout_func_name in [
+                (self.map_params.worker_init_timeout, self._worker_comms.has_worker_init_timed_out, 'worker_init'),
+                (self.map_params.task_timeout, self._worker_comms.has_worker_task_timed_out, 'task')]:
+            if timeout_var is not None:
+                for worker_id in range(self.pool_params.n_jobs):
+                    if has_timed_out_func(worker_id, timeout_var):
+                        # We need to add an exception if we're using the progress bar handler
+                        err_msg = f"Worker-{worker_id} {timeout_func_name} timed out"
+                        if self.map_params.progress_bar:
+                            self._worker_comms.add_exception(TimeoutError, err_msg)
+                        self.terminate()
+                        raise TimeoutError(err_msg)
 
         return obtained_results
 
@@ -226,6 +239,24 @@ class WorkerPool:
         """
         return self._exit_results
 
+    def _get_exit_result_from_worker(self, worker_id) -> None:
+        """
+        Obtains exit results from a single worker. When timeout is reached, it will terminate and raise
+
+        :param worker_id: Worker ID
+        """
+        try:
+            self._exit_results.append(
+                self._worker_comms.get_exit_results(worker_id, self.map_params.worker_exit_timeout)
+            )
+        except TimeoutError:
+            # We need to add an exception if we're using the progress bar handler
+            err_msg = f"Worker-{worker_id} worker_exit timed out"
+            if self.map_params.progress_bar:
+                self._worker_comms.add_exception(TimeoutError, err_msg)
+            self.terminate()
+            raise TimeoutError(err_msg)
+
     def __enter__(self) -> 'WorkerPool':
         """
         Enable the use of the ``with`` statement.
@@ -243,7 +274,9 @@ class WorkerPool:
             max_tasks_active: Optional[int] = None, chunk_size: Optional[int] = None, n_splits: Optional[int] = None,
             worker_lifespan: Optional[int] = None, progress_bar: bool = False, progress_bar_position: int = 0,
             concatenate_numpy_output: bool = True, enable_insights: Optional[bool] = None,
-            worker_init: Optional[Callable] = None, worker_exit: Optional[Callable] = None) -> Any:
+            worker_init: Optional[Callable] = None, worker_exit: Optional[Callable] = None,
+            task_timeout: Optional[float] = None, worker_init_timeout: Optional[float] = None,
+            worker_exit_timeout: Optional[float] = None) -> Any:
         """
         Same as ``multiprocessing.map()``. Also allows a user to set the maximum number of tasks available in the queue.
         Note that this function can be slower than the unordered version.
@@ -281,6 +314,13 @@ class WorkerPool:
             through :obj:`mpire.WorkerPool.get_exit_results`. When passing on the worker ID the function should receive
             the worker ID as its first argument. If shared objects are provided the function should receive those as the
             next argument. If the worker state has been enabled it should receive a state variable as the next argument
+        :param task_timeout: Timeout in seconds for a single task. When the timeout is exceeded, MPIRE will raise a
+            ``TimeoutError``. Use ``None`` to disable (default). Note: the timeout doesn't apply to ``worker_init`` and
+            ``worker_exit`` functions, use `worker_init_timeout` and `worker_exit_timeout` for that, respectively
+        :param worker_init_timeout: Timeout in seconds for the ``worker_init`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
+        :param worker_exit_timeout: Timeout in seconds for the ``worker_exit`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
         :return: List with ordered results
         """
         # Notify workers to keep order in mind
@@ -297,7 +337,8 @@ class WorkerPool:
             iterable_len = len(iterable_of_args)
         results = self.map_unordered(func, ((args_idx, args) for args_idx, args in enumerate(iterable_of_args)),
                                      iterable_len, max_tasks_active, chunk_size, n_splits, worker_lifespan,
-                                     progress_bar, progress_bar_position, enable_insights, worker_init, worker_exit)
+                                     progress_bar, progress_bar_position, enable_insights, worker_init, worker_exit,
+                                     task_timeout, worker_init_timeout, worker_exit_timeout)
 
         # Notify workers to forget about order
         self._worker_comms.clear_keep_order()
@@ -314,7 +355,9 @@ class WorkerPool:
                       chunk_size: Optional[int] = None, n_splits: Optional[int] = None,
                       worker_lifespan: Optional[int] = None, progress_bar: bool = False,
                       progress_bar_position: int = 0, enable_insights: Optional[bool] = None,
-                      worker_init: Optional[Callable] = None, worker_exit: Optional[Callable] = None) -> Any:
+                      worker_init: Optional[Callable] = None, worker_exit: Optional[Callable] = None,
+                      task_timeout: Optional[float] = None, worker_init_timeout: Optional[float] = None,
+                      worker_exit_timeout: Optional[float] = None) -> Any:
         """
         Same as ``multiprocessing.map()``, but unordered. Also allows a user to set the maximum number of tasks
         available in the queue.
@@ -351,19 +394,28 @@ class WorkerPool:
             through :obj:`mpire.WorkerPool.get_exit_results`. When passing on the worker ID the function should receive
             the worker ID as its first argument. If shared objects are provided the function should receive those as the
             next argument. If the worker state has been enabled it should receive a state variable as the next argument
+        :param task_timeout: Timeout in seconds for a single task. When the timeout is exceeded, MPIRE will raise a
+            ``TimeoutError``. Use ``None`` to disable (default). Note: the timeout doesn't apply to ``worker_init`` and
+            ``worker_exit`` functions, use `worker_init_timeout` and `worker_exit_timeout` for that, respectively
+        :param worker_init_timeout: Timeout in seconds for the ``worker_init`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
+        :param worker_exit_timeout: Timeout in seconds for the ``worker_exit`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
         :return: List with unordered results
         """
         # Simply call imap and cast it to a list. This make sure all elements are there before returning
         return list(self.imap_unordered(func, iterable_of_args, iterable_len, max_tasks_active, chunk_size,
                                         n_splits, worker_lifespan, progress_bar, progress_bar_position,
-                                        enable_insights, worker_init, worker_exit))
+                                        enable_insights, worker_init, worker_exit, task_timeout, worker_init_timeout,
+                                        worker_exit_timeout))
 
     def imap(self, func: Callable, iterable_of_args: Union[Sized, Iterable], iterable_len: Optional[int] = None,
              max_tasks_active: Optional[int] = None, chunk_size: Optional[int] = None, n_splits: Optional[int] = None,
              worker_lifespan: Optional[int] = None, progress_bar: bool = False,
              progress_bar_position: int = 0, enable_insights: Optional[bool] = None,
-             worker_init: Optional[Callable] = None,
-             worker_exit: Optional[Callable] = None) -> Generator[Any, None, None]:
+             worker_init: Optional[Callable] = None, worker_exit: Optional[Callable] = None,
+             task_timeout: Optional[float] = None, worker_init_timeout: Optional[float] = None,
+             worker_exit_timeout: Optional[float] = None) -> Generator[Any, None, None]:
         """
         Same as ``multiprocessing.imap_unordered()``, but ordered. Also allows a user to set the maximum number of
         tasks available in the queue.
@@ -400,6 +452,13 @@ class WorkerPool:
             through :obj:`mpire.WorkerPool.get_exit_results`. When passing on the worker ID the function should receive
             the worker ID as its first argument. If shared objects are provided the function should receive those as the
             next argument. If the worker state has been enabled it should receive a state variable as the next argument
+        :param task_timeout: Timeout in seconds for a single task. When the timeout is exceeded, MPIRE will raise a
+            ``TimeoutError``. Use ``None`` to disable (default). Note: the timeout doesn't apply to ``worker_init`` and
+            ``worker_exit`` functions, use `worker_init_timeout` and `worker_exit_timeout` for that, respectively
+        :param worker_init_timeout: Timeout in seconds for the ``worker_init`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
+        :param worker_exit_timeout: Timeout in seconds for the ``worker_exit`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
         :return: Generator yielding ordered results
         """
         # Notify workers to keep order in mind
@@ -420,7 +479,8 @@ class WorkerPool:
                                                              in enumerate(iterable_of_args)), iterable_len,
                                                       max_tasks_active, chunk_size, n_splits, worker_lifespan,
                                                       progress_bar, progress_bar_position, enable_insights, worker_init,
-                                                      worker_exit):
+                                                      worker_exit, task_timeout, worker_init_timeout,
+                                                      worker_exit_timeout):
 
             # Check if the next one(s) to return is/are temporarily stored. We use a while-true block with dict.pop() to
             # keep the temporary store as small as possible
@@ -451,8 +511,9 @@ class WorkerPool:
                        chunk_size: Optional[int] = None, n_splits: Optional[int] = None,
                        worker_lifespan: Optional[int] = None, progress_bar: bool = False,
                        progress_bar_position: int = 0, enable_insights: Optional[bool] = None,
-                       worker_init: Optional[Callable] = None,
-                       worker_exit: Optional[Callable] = None) -> Generator[Any, None, None]:
+                       worker_init: Optional[Callable] = None, worker_exit: Optional[Callable] = None,
+                       task_timeout: Optional[float] = None, worker_init_timeout: Optional[float] = None,
+                       worker_exit_timeout: Optional[float] = None) -> Generator[Any, None, None]:
         """
         Same as ``multiprocessing.imap_unordered()``. Also allows a user to set the maximum number of tasks available in
         the queue.
@@ -489,6 +550,13 @@ class WorkerPool:
             through :obj:`mpire.WorkerPool.get_exit_results`. When passing on the worker ID the function should receive
             the worker ID as its first argument. If shared objects are provided the function should receive those as the
             next argument. If the worker state has been enabled it should receive a state variable as the next argument
+        :param task_timeout: Timeout in seconds for a single task. When the timeout is exceeded, MPIRE will raise a
+            ``TimeoutError``. Use ``None`` to disable (default). Note: the timeout doesn't apply to ``worker_init`` and
+            ``worker_exit`` functions, use `worker_init_timeout` and `worker_exit_timeout` for that, respectively
+        :param worker_init_timeout: Timeout in seconds for the ``worker_init`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
+        :param worker_exit_timeout: Timeout in seconds for the ``worker_exit`` function. When the timeout is exceeded,
+            MPIRE will raise a ``TimeoutError``. Use ``None`` to disable (default).
         :return: Generator yielding unordered results
         """
         # If we're dealing with numpy arrays, we have to chunk them here already
@@ -504,9 +572,10 @@ class WorkerPool:
         # modified as well
         n_tasks, max_tasks_active, chunk_size, progress_bar = check_map_parameters(
             self.pool_params, iterable_of_args, iterable_len, max_tasks_active, chunk_size, n_splits, worker_lifespan,
-            progress_bar, progress_bar_position
+            progress_bar, progress_bar_position, task_timeout, worker_init_timeout, worker_exit_timeout
         )
-        new_map_params = WorkerMapParams(func, worker_init, worker_exit, worker_lifespan)
+        new_map_params = WorkerMapParams(func, worker_init, worker_exit, worker_lifespan, progress_bar, task_timeout,
+                                         worker_init_timeout, worker_exit_timeout)
 
         # enable_insights is deprecated as a map parameter
         if (self._workers and enable_insights is not None and self.pool_params.enable_insights is not None and
@@ -539,7 +608,7 @@ class WorkerPool:
                 self._worker_comms.add_new_map_params(new_map_params)
             if not self._workers:
                 self.map_params = new_map_params
-                self._start_workers(progress_bar)
+                self._start_workers()
 
             # Create progress bar handler, which receives progress updates from the workers and updates the progress bar
             # accordingly
@@ -623,7 +692,7 @@ class WorkerPool:
             if not self._worker_comms.exception_thrown():
                 keyboard_interrupt = True
                 self._worker_comms.signal_exception_thrown()
-                if self._worker_comms.has_progress_bar() and progress_bar_handler is not None:
+                if self.map_params.progress_bar and progress_bar_handler is not None:
                     self._worker_comms.add_exception(KeyboardInterrupt, "KeyboardInterrupt")
 
         # When we're not dealing with a KeyboardInterrupt, obtain the exception thrown by a worker
@@ -696,7 +765,10 @@ class WorkerPool:
             # Obtain results from the exit results queues (should be done before joining the workers)
             if not keep_alive and self.map_params.worker_exit:
                 logger.debug("Obtaining exit results")
-                self._exit_results.extend(self._worker_comms.get_exit_results_all_workers())
+                for worker_id in range(self.pool_params.n_jobs):
+                    self._get_exit_result_from_worker(worker_id)
+                    if self._worker_comms.exception_thrown():
+                        break
                 self._worker_comms.signal_all_exit_results_obtained()
                 logger.debug("Done obtaining exit results")